requirements.txt
python-telegram-bot==13.7
python-dotenv==1.0.0
requests==2.31.0
# Pin specific versions for compatibility
numpy==1.24.3
huggingface-hub==0.16.4
transformers==4.30.2
sentence-transformers==2.2.2
chromadb==0.4.13